Move missing methods from page onto frame. Closes #113
