+++ lib/FrameManager.js
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+let fs = require('fs');
 let EventEmitter = require('events');
 let helper = require('./helper');
 
@@ -157,17 +158,39 @@ class FrameManager extends EventEmitter {
 
   /**
    * @param {!Frame} frame
-   * @param {string} expression
-   * @return {!Promise<(!Object|undefined)>}
+   * @return {number}
    */
-  async _evaluateOnFrame(frame, expression) {
+  _contextIdForFrame(frame) {
     let contextId = undefined;
     if (frame !== this._mainFrame) {
       contextId = this._frameIdToExecutionContextId.get(frame._id);
       console.assert(contextId, 'Frame does not have default context to evaluate in!');
     }
+    return contextId;
+  }
+
+  /**
+   * @param {!Frame} frame
+   * @param {string} expression
+   * @return {!Promise<(!Object|undefined)>}
+   */
+  async _evaluateOnFrame(frame, expression) {
+    let contextId = this._contextIdForFrame(frame);
     expression = `Promise.resolve(${expression})`;
     let { exceptionDetails, result: remoteObject }  = await this._client.send('Runtime.evaluate', { expression, contextId, returnByValue: false, awaitPromise: true });
+    if (exceptionDetails)
+      throw new Error('Evaluation failed: ' + helper.getExceptionMessage(exceptionDetails));
+    return await helper.serializeRemoteObject(this._client, remoteObject);
+  }
+
+  /**
+   * @param {!Frame} frame
+   * @param {string} expression
+   * @return {!Promise<(!Object|undefined)>}
+   */
+  async _rawEvaluateOnFrame(frame, expression) {
+    let contextId = this._contextIdForFrame(frame);
+    let { exceptionDetails, result: remoteObject }  = await this._client.send('Runtime.evaluate', { expression, contextId, returnByValue: false });
     if (exceptionDetails)
       throw new Error('Evaluation failed: ' + helper.getExceptionMessage(exceptionDetails));
     return await helper.serializeRemoteObject(this._client, remoteObject);
@@ -252,6 +275,40 @@ class Frame {
   }
 
   /**
+   * @param {string} filePath
+   * @return {!Promise}
+   */
+  async injectFile(filePath) {
+    let contents = await new Promise((resolve, reject) => {
+      fs.readFile(filePath, 'utf8', (err, data) => {
+        if (err) return reject(err);
+        resolve(data);
+      });
+    });
+    contents += `//# sourceURL=` + filePath.replace(/\n/g,'');
+    return this._frameManager._rawEvaluateOnFrame(this, contents);
+  }
+
+  /**
+   * @param {string} url
+   * @return {!Promise}
+   */
+  async addScriptTag(url) {
+    return this.evaluate(addScriptTag, url);
+
+    /**
+     * @param {string} url
+     */
+    function addScriptTag(url) {
+      let script = document.createElement('script');
+      script.src = url;
+      let promise = new Promise(x => script.onload = x);
+      document.head.appendChild(script);
+      return promise;
+    }
+  }
+
+  /**
    * @param {(string|number)} target
    * @param {!Object=} options
    * @return {!Promise}
@@ -314,6 +371,13 @@ class Frame {
   }
 
   /**
+   * @return {!Promise<string>}
+   */
+  async title() {
+    return this.evaluate(() =>  document.title);
+  }
+
+  /**
    * @param {string} selector
    * @return {!Promise}
    */
@@ -344,6 +408,22 @@ class Frame {
     await this._frameManager._mouse.press(options);
     // This is a hack for now, to make clicking less race-prone
     await this.evaluate(() => new Promise(f => requestAnimationFrame(f)));
+  }
+
+  /**
+   * @param {string} selector
+   * @return {!Promise}
+   */
+  async focus(selector) {
+    let success = await this.evaluate(selector => {
+      let node = document.querySelector(selector);
+      if (!node)
+        return false;
+      node.focus();
+      return true;
+    }, selector);
+    if (!success)
+      throw new Error('No node found for selector: ' + selector);
   }
 
   /**
