+++ lib/Page.js
@@ -121,18 +121,7 @@ class Page extends EventEmitter {
    * @return {!Promise}
    */
   async addScriptTag(url) {
-    return this.evaluate(addScriptTag, url);
-
-    /**
-       * @param {string} url
-       */
-    function addScriptTag(url) {
-      let script = document.createElement('script');
-      script.src = url;
-      let promise = new Promise(x => script.onload = x);
-      document.head.appendChild(script);
-      return promise;
-    }
+    return this.mainFrame().addScriptTag(url);
   }
 
   /**
@@ -140,14 +129,7 @@ class Page extends EventEmitter {
    * @return {!Promise}
    */
   async injectFile(filePath) {
-    let contents = await new Promise((resolve, reject) => {
-      fs.readFile(filePath, 'utf8', (err, data) => {
-        if (err) return reject(err);
-        resolve(data);
-      });
-    });
-    contents += `//# sourceURL=` + filePath.replace(/\n/g,'');
-    return this._client.send('Runtime.evaluate', { expression: contents, returnByValue: true });
+    return this.mainFrame().injectFile(filePath);
   }
 
   /**
@@ -507,7 +489,7 @@ class Page extends EventEmitter {
    * @return {!Promise<string>}
    */
   async title() {
-    return this.evaluate(() =>  document.title);
+    return this.mainFrame().title();
   }
 
   /**
@@ -561,15 +543,7 @@ class Page extends EventEmitter {
    * @return {!Promise}
    */
   async focus(selector) {
-    let success = await this.evaluate(selector => {
-      let node = document.querySelector(selector);
-      if (!node)
-        return false;
-      node.focus();
-      return true;
-    }, selector);
-    if (!success)
-      throw new Error('No node found for selector: ' + selector);
+    return this.mainFrame().focus(selector);
   }
 
   /**
